code/18 - Image to ASCII Art/requirements.txt
Pillow==9.5.0
numpy==1.24.3
